Implement direct_tcpip example from libssh2 website
This example establishes a reverse TCP/IP tunnel that the client can
use to connect with another system through the server.

The example is prepared to tunnel a SSH session through the session
created by the example. If everything is setup you can use PuTTY or
a similar tool to establish a connection through the tunnel.

For more details see the original example:
https://libssh2.org/examples/direct_tcpip.html

Clear Errors (Backport).vi was taken from LabVIEW 2019 and saved
back for LabVIEW 2013. The code remains unchanged and can be replaced
with the native function when needed.