Refactor LIBSSH2_SESSION* into Session refnum
All functions that require a session pointer now require a session
refnum that serves as wrapper for the session pointer.

Unfortunately it is not possible to directly cast a pointer-sized
integer into a refnum because pointer-sized integers are 64-bit and
refnums are only 32-bit. Instead it uses a data value reference to
the pointer value.

Also included is a function that callers can use to determine if the
session refnum is valid (Not A Session.vi).