Refactor libssh2_knownhost* into Host refnum
This refnum is a bit special because it represents a struct.

Unfortunately we cannot simply return a cluster for the struct because
there are functions that require the raw pointer in order to function.
Because of this, callers must explicitly extract the cluster from
the Host refnum using 'Read Host.vi' and free memory using
'Free Host.vi'.