libssh2_hostkey_hash: Fix return value is not null-terminated string
When requesting the hostkey hash, we define the hash type, which
has a defined length depending on the chosen type. Since we can only
know the length of hash types that we are familiar with, the hash
type must be specified as enum. That way, if new types are added in
the future, we are also forced to specify the expected hash length.