Allow users to specify the library to use
Users can now specify their own library in libssh2_init(). By default
it will use the library that is included with the library. The path
is reset when calling libssh2_exit() so that libssh2_init() must be
called again before using any other function.

This also adds a new function libssh2_version() to return the version
number of libssh2.