Use polymorphic VI to include library in executables
Since the path to the library can be specified by the user, it is no
longer included when building an application. In order to have the
library included with the application again, we need a polymorphic
VI that provides an instance to statically link to the library.
Only when the user specifies a path, we use the other istance that
links via the path. The other VIs don't need to change as they use
the path defined in the init function.