libssh2: Fix access violation reading known host data
Although the implementation appears to be working, it really isn't.
That's because I'm developing in a 32-bit environment and attempt to
read 64-bit pointers, which results in overflows reading the pointers
and accessing random memory as a result of that.

Fixed by reading 32-bit pointers on 32-bit environments and 64-bit
pointers on 64-bit environments.

We also don't need to initialize our own pointers as libssh2 returns
their own.