libssh2: Fix remnant value due to uninitialized CLFN parameters
As it turns out, the CLFN does not initialize unwired parameters on
every call. Instead, the same memory location is reused, which can
result in remnant values being returned when the underlying function
did not change its content.

This could be observed predominantly in the sftpdir() example, where
the final directory listing was listed twice. After initializing the
parameter, the duplicate entry does not occur anymore.