libssh2: Add libssh2_sftp_write()
This also adds an example that writes content to a file over SFTP.

libssh2 writes files in chunks of 30k bytes and returns how many bytes
have been acknowledged by the server so far. This value can be zero in
non-blocking mode. Callers must assume that the entire buffer was written
and continue with normal operation. There is no way to tell when all
bytes are acknowledged by the server.