libssh2: Add libssh2_sftp_mkdir()
Technically, libssh2_sftp_mkdir is a macro for libssh2_sftp_mkdir_ex
but in our case we don't need the *_ex function because the only
difference is how path_len is provided. In the case of *_ex, it is
provided by the caller. libssh2_sftp_mkdir, on the other hand, uses
the strlen() function to determine path_len. We use the same approach
for all strings anyways, which means that both libssh2_sftp_mkdir and
libssh2_sftp_mkdir_ex would have the same implementation. For this
reason libssh2_sftp_mkdir_ex is omitted.