tests: Test all libssh2_* functions for error 1097
Instead of writing a test function for each function in the library,
we can simply iterate and execute all functions and check if the
returned error code is 1097.

Note that some functions may fail for different reasons. This test
only checks that error 1097 does not occur, which would indicate a
possible memory corruption.

Some functions are excluded from the test as they have side-effects
(e.g. libssh2_init* and libssh2_exit). The test verifies that all
excluded functions are accounted for. This is to make sure that any
renamed or removed files are detected.