libssh2: Add libssh2_session_callback_set()
This also changes the implementation for libssh2_session_init_ex() to
accept a static VI Refnum for the 'send' callback. The implementation
of libssh2_session_callback_set() is limited to the 'send' callback
type for now (more callbacks will follow in the future).

This also adds new function to lvssh2_extensions.dll, so that callbacks
can be handled in LabVIEW using event registration on user events that
are managed as part of the session.

When the session is closed, the associated user events and event
registrations are also deleted.

The implementation is currently limited to 32-bit. For 64-bit we need
different functions to exchange data with the underlying extension library.

The session_callback example utilizes a custom 'send' function which
depends on winsock2 to send the data on the provided socket. The library
itself only contains the template VI which defines the VI pattern for the
static VI Refnum. Callers are responsible for providing appropriate
implementations. By default, the libssh2 implementations are used.