libssh2: Add libssh2_session_abstract()
This also changes the implementation of libssh2_session_init_ex() to
store the provided 'abstract' (VI Refnums) in the session handle so
that it can be retrieved for libssh2_session_abstract().

Note that libssh2_session_abstract() is technically a no-op because
it simply returns the VI Refnums stored in the session handle. It can
however, be extended in the future, so that the registered callbacks
are swapped-out in place (e.g., to dynamically replace a function).