libssh2: Always write entire buffer when reading/writing files
libssh2_sftp_read() and libssh2_sftp_write() return immediately when
the first block is send to the remote SSH server. It expects that the
caller provides the same buffer pointer and data length on the next
call so that the next block can be send.

In LabVIEW, this has the underiable result that any block of data with
more than 30000 bytes is truncated as we don't have a pointer but a
string value. To simplify the implementation for callers of lvssh2,
the string is internally copied into a buffer and send in a loop. This
does work for any size of data but will block execution until all blocks
are sent.