libssh2: Fix softlock due to NULL pointer
libssh2_knownhost_writeline() may softlock when provided with a NULL
pointer for the known host. The issue was observed in LabVIEW 2019
SP1 (32-bit) although it does not seem to appear in LabVIEW 2023 SP1
(32-bit).

Fixed by adding 'Error if not a Known Host' and verifying the known
host before calling external code in libssh2_knownhost_writeline()
and 'Read Known Host Data'.

This also improves the implementation of the examples to first read
all known hosts and then extract information from them. This is
necessary because the final known host returned by libssh2_knownhost_get()
may be invalid.