Toolchain/Test: Report execution errors to the CLI
In some situations it appears that an execution error prevents the
test from succeeding, resultin in no test results being exported.
Azure Pipelines is configured to ignore missing results so that it
appears as successful, which is clearly not the case. By reporting
execution errors to the CLI, the actual error should be reported.