libssh2: Update VI titles and descriptions
lvssh2-descriptions.xml contains the titles and descriptions for all
public VIs and is applied with 'Update Descriptions.vi'.

Instead of manually updating titles and descriptions for each VI
individually, this is much easier as it also enables the use of more
powerful text editors (e.g., with Copilot support).

All libssh2 functions have a human-readable name and the libssh2
function name in parenthesis. This is done to make it easier to find
functions in quick-drop.