libssh2: Change the naming for known host(s) functions
This changes the naming of 'known host' to 'host key', following the
convention in RFC4252:

- 'known host' -> 'host key'
- 'known hosts' -> 'host key database'

Note that the host key database represents the 'known_hosts' file that
is typically used by OpenSSH but is not restricted to it (e.g., a caller
may want to implement 'blocked_hosts' with the same fundamental stucture).

Function names from libssh2 are left unchanged.