libssh2: Do not report errors for uninitialized refnums
This prevents an error when initializing a session for a platform that
is not fully implemented. In this case, the event refnum and callbacks
are never initialized, which results in an error when releasing the
session. Since the functionality of other libssh2 functions are not
affected by missing callbacks, no error should be emitted.