libssh2: Fix issues related to library name resolution
VIPM replaces all wildcard library paths with explicit paths during
the build process. For example, "libssh2**.*" turns into "libssh2.dll"
when build with 32-bit LabVIEW on Windows.

This results in broken VIs when installing the library on different
bitness or plattform. This is a known issue that must be resolved with
a custom post-install action[1].

[1] https://support.vipm.io/hc/en-us/articles/360053941572-VI-Package-Builder-removes-cross-platform-wildcards-from-Call-Library-Function-nodes-during-the-build-process