tests: Add docker compose file for linuxserver/openssh-server
We currently use sshd of the host system for integration tests, which
is problematic the environment is not sandboxed and provides no guard
for isolation against malicious code. It also does not work on the
build agent because, when the agent runs as a service (recommended),
it is not possible to login with the user account (service account is
not a real user).

With the docker container, we can isolate the sshd server into a re-
usable component that will run on any system (even on linux should we
want to port in the future).

This commit adds the base configuration for setting up the server and
providing the sshd_config file (generated/updated by the container on
startup).

[1] https://hub.docker.com/r/linuxserver/openssh-server