tests: Fix unable to exchange encryption keys
There appears to be a bug in libssh2 when using the WINCNG crypto
backend, which results in error -8 "Unable to exchange encryption
keys" during key exchange. The error happens during diffie-hellman
key-pair generation ("dh key pair generation failed") with method
'diffie-hellman-group-exchange-sha256'.

This appears to be a known issue that may be addressed in future
versions of libssh2: https://github.com/libssh2/libssh2/issues/1370

The workaround is to disable the offending method in sshd_config.