libssh2: Fix incorrect pointer-dereference in libssh2_session_abstract
The value returned by libssh2_session_abstract() is **abstract but
we handle it as *abstract. Fixed by first dereferencing the pointer
to *abstract before dereferencing the actual value.

This currently does not affect any users because the value is not
exposed by the VI. Nonetheless, it could result in a crash on 32-bit
platforms when attempting to read 8 bytes from a 4-byte address pointer.