libssh2: Add callback input to libssh2_callback_set()
This input allows callers to specify their own callback function. The
behavior of the function changes depending on whether lvssh2 extensions
are enabled or disabled.

When the extensions are enabled, setting callback to 0 (default) will
use the default built-in callback function.

When the extensions are disabled, setting callback to 0 (default) will
disable the callback.

There is currently no way to disable the callback when using extensions.