Fix crashes in libssh2_userauth_keyboard_interactive_ex()
These crashes occurred due to alignment issues between LabVIEW and
libssh2_extensions. Fixed by addressing the alignment issues on the
LabVIEW side:
- `prompts` changed from U32 to U64 (pointer-sized integer)
- padding added for `echo` so that the structure is correctly aligned.

To address a heap corruption issue in LabVIEW, the implementation for
the handler function was changed so that the `responses` handle is also
provided to the callback function. The corresponding helper function
`lvssh2_userauth_keyboard_interactive_add_response` was changed to use
the provided `responses` handle and `index` to add the response. This
was previously done with global variables that are no longer necessary.

Side note: There is another issue related to heap corruption that isn't
resolved by this commit. It does, however, not result in a crash and
will be addressed in a future commit.