libssh2/libssh2_userauth_keyboard_interactive_ex: Register callback function in libssh2_session_init_ex()
This removes the 'response' input from
libssh2_userauth_keyboard_interactive_ex() and adds a new parameter
'keyboard-interactive response' to the abstract data provided to
libssh2_session_init_ex()

This is done so the callback can be registered to the abstract type
that is provided to the callback handler, which makes keyboard-
interactive authentication thread-safe as the operation is executed
on the session object and not through a global variable.

This makes the setup of keyboard-interactive authentication slightly
more awkward as the callback handler must be registered during session
initialization. However, it avoids any disambiguation when authenticating
in multiple parallel threads.