libssh2: Fix infinite loop when enabling callback without handlers
When libssh2_session_callback_set() is used to activate a callback
for which no handler was registered in libssh2_session_init_ex(),
this can result in an infinite loop because the user event from
lvssh2 is always registered automatically. When the callback is
activated, it always returns default values because the VI refnum
is undefined.

This changes the implementation so that the LabVIEW user event is
only created when a valid VI refnum is provided. The extension
library returns an error if the user event does not exist, which
effectively breaks any infinite loops.