libssh2: Enable reentrant execution on (almost) all functions
Callers are currently responsible to ensure that handles are accessed
sequentially. Otherwise, the behavior is undefined.

libssh2_init() is not reentrant because the function is not thread-
safe according to the libssh2 documentation. Note that this also
applies to libssh2_session_init*(), which calls libssh2_init()
internally as convenience function. Though, this is not an issue
in lvssh2 because libssh2_session_init*() does not work unless
libssh2_init() has been called (library path must be initialized).

There are no automated tests to validate correct behavior because it
is almost impossible to validate correct behavior of parallel threads
as race-conditions are non-deterministic. The best we can do is add
locks to prevent users from accidentally corrupting memory by calling
multiple functions on a single object but this is not the scope of
this commit.