libssh2: Add semaphore to session handle
This adds a semaphore to the session handle to prevent callers from
executing multiple functions on one session in parallel, which
could result in corrupted memory and undefined behaviour.

The semaphore is session-specific to allow multiple sessions to be
used at once. It is currently not shared with channels related to a
session.