libssh2: Add semaphore to the channel handle
This adds a semaphore to the channel handle to prevent concurrent
access to the channel. Since channels in libssh2 utilize the TCP
connection of the session it was created from, the semaphore used
for the channel is shared from the session. This ensures that all
functions utilizing the session do not conflict with another.