libssh2_channel_process_startup: Fix unexpected bytes remain after decoding
When libssh2_channel_process_startup is called with an empty message,
libssh2 only checks whether the message pointer is not NULL and does
not check if message_len is zero. This results in extra bytes being
added to the message, resulting in an error on the server side.

This is fixed by providing NULL for message when the message length
is zero.