libssh2_sftp_read/write: Change buffer type from string to byte array
This changes the data type of the 'buffer' terminals on the SFTP read
and write functions from string to byte array as these function can be
used to read/write arbitrary binary data and are not limited to strings.

This is a breaking change. Callers must update their code to convert
from/to string using the 'Byte Array To String' and 'String To Byte
Array' functions, respectively.

The VI icons were updated to reflect this change and now have a 'binary'
glyph.

All examples and tests were updated to convert from/to string as needed.