CMakeLists: Fix discovery of 'cintools' independent of bitness
Both 32-bit and 64-bit builds can utilize the same 'cintools' files,
regardless if the build is for 32 or 64 bit. We currently discover
the files base on the bitness, which does not work unless the host
has both 32 and 64 bit versions of LabVIEW installed.

This fixes the issue by always considering both 32 and 64 bit folders
when searching for 'cintools'.