CMakeLists: Introduce LIBSSH2_COMMIT_HASH
LIBSSH2_COMMIT_HASH specifies the commit hash of the libssh2 repository
used for the build process. It also accepts branch names and tags.
However, the use of commit hashes is strongly recommended to avoid
pulling malicious code.

The default value is set to the commit hash of libssh2-1.11.0

This also updates azure-pipelines.yml to explicitly specify the commit
hash to use for the pipeline.