CMakeLists: Build libssh2 without deprecated APIs
This is a new feature in libssh2 1.11.1 which excludes deprecated APIs
when compiling the project. Affected APIs:

- libssh2_banner_set
- libssh2_session_startup
- libssh2_channel_receive_window_adjust
- libssh2_channel_handle_extended_data
- libssh2_scp_recv

None of these APIs have bindings in lvssh2 so it is safe to exclude
them.

Note that LIBSSH2_NO_DEPRECATED does not actually remove all deprecated
APIs but only those that are deprecated for a long time (e.g.,
libssh2_session_callback_set is deprecated but not removed)