libssh2: Mark libssh2_session_callback_set.vi as deprecated
libssh2_session_callback_set() is deprecated in libssh2-1.11.1

This adds a deprecation warning to the VI description and updates the
icon to stand out on the block diagram. The error cluster returns a
warning with a brief explanation when the VI is called.

Note that libssh2_session_callback_set(), while deprecated, is still
available in libssh2-1.11.1.