libssh2: Add libssh2_session_callback_set2.vi
libssh2_session_callback_set2() was added in libssh2-1.11.1 as a
replacement for libssh2_session_callback_set(), which is not type safe.

This also updates the extensions library to utilize the new
libssh2_cb_generic type and replaces the old function in all callers and
the palette.

Note that the implementation between *set and *set2 in LabVIEW are
practically identical because all instances of libssh2_cb_generic* are
simply exposed as pointer-sized integers.