libssh2: Add libssh2_sftp_posix_rename_ex.vi
libssh2_sftp_posix_rename_ex() was added in libssh2-1.11.1. It provides
POSIX semantics for rename operations. This means that the file appears
to be moved to any services monitoring for file changes [1]. It is not
supported by all SSH servers.

This also adds more descriptive descriptions to all rename functions.

[1]: https://github.com/PowerShell/openssh-portable/blob/v8.1.0.0/PROTOCOL#L380..L395