tests: Properly close the channel before reading exit status
This fixes unstable tests for libssh2_channel_get_exit_signal() and
libssh2_channel_get_exit_status(). These functions only return valid
values when the corresponding channel is properly closed from both
sides. Previously, the status entirely depended on the performance of
the SSH server; if it was fast enough to respond with the exit status
before it was read by the tests, the test would pass. Otherwise, they
would fail.

This is fixed by properly closing the channel, sending EOF and waiting
for the corresponding responses.