azure-pipelines: Use strategy to configure build configurations
The build pipeline currently makes no assumption on the build agent and
simply builds for both x86 and x64 targets. It is then up to the agent
to use whatever it can handle. The problem with this approach is that it
is not clear to the viewer that only the x86 or x64 target has been
tested (in fact, only the x86 target is tested).

This changes the pipeline to specify the target and configuration
explicitly.