build.sh: replace release.sh, add platform support for Windows and Linux
This replaces release.sh and adds platform support to build lvssh2 on
Windows and Linux. It supports a configuration option to specify whether
to make a Debug or Release build. For Debug builds, the GitHub
repository is used (this is to avoid the signature check). The Release
build works as previously.