CMakeLists: Ensure compatibility with cmake 3.22
This replaces CMAKE_HOST_WIN32 and CMAKE_HOST_LINUX with string
comparisons for CMAKE_HOST_SYSTEM_NAME. The change is necessary because
CMAKE_HOST_LINUX was added in cmake 3.25, which is not included with
Ubuntu 22.04 by default. Since Ubuntu 22.04 is the latest version
officially supported by LabVIEW for linux, the included version should
be supported.