libssh2: Generate platform-specific default library path
The search path for shared libraries in LabVIEW depends on the current
platform [1]. On Windows, the current VI's path is included but on other
platforms it isn't. This results in errors when attempting to call the
library with only the filename "libssh2**.*" because the library is not
in the default search paths (both in the IDE and at runtime).

This adds a function to build the path to the shared library depending
on the current platform and environment (currently works for Windows and
Linux). On Windows, the default filename is used. On linux, the search
path is changed to an absolute path, either relative to the current VI's
path (in the IDE) or the data directory (runtime).

The alternative is to register libssh2 and the extensions library into
the standard search paths for each platform, which is error prone and
may not work with the available tooling (e.g., through VIPM).

Another alternative is to distribute the shared libraries independent of
this project, which adds unecessary and potentially complicated tasks
that should be avoided.

It would also be possible to remove the ability to specify a custom
libssh2 library, in which case the library path is gracefully resolved
by LabVIEW (the library path is stored relative by LabVIEW) but this
removes the ability to link against a custom/specific version of libssh2
that is a useful feature (e.g., for debugging purposes).

[1] https://knowledge.ni.com/KnowledgeArticleDetails?id=kA00Z0000019XgCSAU