Bump source version to LabVIEW 2017
This is a mass-compile for LabVIEW 2017 to allow development on newer
versions of LabVIEW; more specifically LabVIEW 2024 Q3, which can save
files for previous versions (17.0 or higher) [1]

This is especially important for Linux hosts, which are tied to certain
versions of LabVIEW [2]. Here, again, the lowest available version is
LabVIEW 2017. By using a source version of 17.0 or higher, Linux hosts
can utilize LabVIEW 2024 and newer to make changes without introducing
breaking changes.

[1] https://labviewwiki.org/wiki/LabVIEW_2024_Q3
[2] https://www.ni.com/en/support/documentation/compatibility/20/labview-and-linux-os-compatibility.html