libssh2: Use conditional symbol TARGET_TYPE instead of OS
The OS conditional symbol is only available for VIs in projects, which
means that any VI that runs outside a project will use the default case.
This can result in undefined behavior when the caller is not aware of
this.

Using TARGET_TYPE avoids this issue as this symbol is also available
outside of projects.