lvssh2.vipb: Add pre-build custom action to update example in README
The example snippet in the README file is broken. This happened because
of incompatible API changes since version 0.1. To ensure that the
example is always functional, this adds a pre-build custom action to
re-generate the snippet for every build.

The pre-build action uses the Code Capture Tool to create the snippet
from examples/ssh2.vi. This means that the snippet does not have the
"border" that the LabVIEW snippet has.