libssh2: Fix segmentation fault in libssh2_session_callback_set*()
This fixes a segmentation fault in libssh2_session_callback_set() and
libssh2_session_callback_set2() when extensions are disabled and the VI
is called with a NULL pointer (0 = use built-in).

This is fixed by adding an additional NULL check to skip setting the
callback when a NULL pointer is provided.