libssh2: Ensure that OpenSSL dependencies are linked
When libssh2 is compiled with OpenSSL, the corresponding libssl-3 and
libcrypto-3 libraries must be included when distributing an application
on Windows. LabVIEW does not automatically include these dependencies
when building an application, which results in a broken build.

For the libraries included with lvssh2, the OpenSSL libraries can be
linked through a trick that suggerates LabVIEW that the library is a
direct dependency without calling it explicitly.

Note: This may break if the LabVIEW compiler ever gets smart enough to
optimize code that is connected to an error constant.