Toolchain: Update post-install step for OpenSSL libraries
With the changes introduced in 39d2f03, the post-install step currently
assumes that any library linked by the project is named using the **.*
schema. But this is not true for the OpenSSL libraries, which use the
'-x64' suffix and thus don't work anymore when the package is installed
by VIPM.

To fix this issue, the post-install step now has a hard-coded set of
libraries to rename.