libssh2: Add helper function to determine SFTP file types
SFTP file types are encoded in the 'permissions' field of the file
attribute. To simplify access to the file types, a new function 'SFTP
File Type' is introduced with similar behavior as the 'File Type'
function of LabVIEW. It converts the permissions values to a simplified
numerical value for the different file types. A corresponding 'SFTP File
Type' type definition exists to define the different values.