tests: Fix subsequent assertion failures in tests with muliple assertions
When a test case has more than one assertion, a failure in the first
assertion will cause the following assertions to fail because we
currently chain assertions with the error wire.

This changes the implementation by using a separate assertion to check
that the error input wire has no error. It also collects the error wires
of all assertions individually to avoid a failure in one assertion to
impact another.