tests: Fix test failure in sftp_readdir_attr
This test fails because of a difference in the file system on the build
agent, which has a block size of 512 instead of 4096. Fixed by asserting
against an array of allowed block sizes [512, 1K, 2K, 4K, 8K].

This also changes the 'Assert Contains' VI into a VIM that utilizes
JSONtext to report the expected value.