Publish code coverage results
This adds a build task to publish code coverage results to the pipeline.
Note that code coverage is reported as "number of lines", which does not
work well with LabVIEW but gives some indicator as to how much code is 
covered.

Code coverage does not include any of the non-LabVIEW source files.