Cache binaries to speed up PR builds
This adds a cache task to restore binaries for libssh2 and OpenSSL so 
long as their commit hash has not changed. The current build time of 
these libraries is approx. 40 minutes, which can be avoided for most PR 
builds as these commit hashes seldomly change.

If the files can be restored from cache, the steps 'Prepare build 
configuration' and 'Build libraries' are skipped.